requirements.txt
torch
torchvision
facenet-pytorch
streamlit
plotly
opencv-python
numpy
pillow